// data/superApp.d.ts
export declare const religiousServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const tourismServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const foodServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const transportServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const commercialServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const propertyServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const rentalServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const courierServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const businessServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const governmentServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const eventServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const petServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const shoppingServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const utilityServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const healthcareServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const emergencyServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const helperServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];
export declare const pilgrimServices: {
    slug: string;
    name: string;
    nameHi: string;
    basePrice: number;
    icon: string;
}[];